Dummy Admin Backend & Introduce finer levels of logging

+++ src/Backend/src/main/java/ie/mid/identityengine/service/StorageService.java
@@ -2,7 +2,10 @@ package ie.mid.identityengine.service;
 
 
 import ie.mid.identityengine.exception.StorageException;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.util.FileSystemUtils;
 
@@ -15,8 +18,13 @@ import java.util.*;
 @Service
 public class StorageService {
 
+    @Value("${mid.uploadPath}")
     private String uploadPathString = "/home/cillian/uploads/";
+    @Value("${mid.eraseOnInit}")
+    private boolean eraseOnInit = true;
+
     private Path uploadPath;
+    private Logger logger = LogManager.getLogger(StorageService.class);
 
     @Autowired
     public StorageService() {
@@ -34,10 +42,13 @@ public class StorageService {
     public String saveData(String data) {
         String name = getUUID();
         String path = uploadPathString + name;
+        logger.debug("Saving data to file " + uploadPathString);
+
         try (Writer writer = new BufferedWriter(new OutputStreamWriter(
                 new FileOutputStream(path), "utf-8"))) {
             writer.write(data);
         } catch (IOException e) {
+            logger.error("Error saving data " + data + " to file " + uploadPathString);
             return null;
         }
         return path;
@@ -48,6 +59,7 @@ public class StorageService {
     }
 
     public String loadData(String filePath) {
+        logger.debug("Loading data from file " + filePath);
         StringBuilder resultStringBuilder = new StringBuilder();
         try (BufferedReader br
                      = new BufferedReader(new InputStreamReader(new FileInputStream(filePath)))) {
@@ -56,19 +68,24 @@ public class StorageService {
                 resultStringBuilder.append(line).append("\n");
             }
         } catch (IOException e) {
+            logger.error("Error loading data from file " + filePath);
             return null;
         }
         return resultStringBuilder.toString();
     }
 
     public void deleteAll() {
-        FileSystemUtils.deleteRecursively(uploadPath.toFile());
+        if (eraseOnInit) {
+            logger.debug("UPLOADED FILES ERASED");
+            FileSystemUtils.deleteRecursively(uploadPath.toFile());
+        }
     }
 
     public void init() {
         try {
             Files.createDirectories(uploadPath);
         } catch (IOException e) {
+            logger.error("Could not initialise storeage at path " + uploadPath);
             throw new StorageException("Could not initialize storage", e);
         }
     }
